+++ changelog.md
@@ -29,12 +29,11 @@ To suggest more, please [open an issue](https://github.com/Sv443/Node-Notifier/i
         - `?waitForResult` to make the API wait for an interaction with the notification
         - `?actions` to have a set of actions the user can select
 - Added dashboard page that displays Node-Notifier's status
-    <!-- - Added notification log to dashboard -->
-    <!-- - Added password protection to dashboard -->
+    TODO: - Added password protection to dashboard
 - Added pm2 wrapper so the service runs in the background
 - Added config file at `./config.js`
 - Added update checker
-    <!-- - Update instructions in dashboard -->
+    - Update instructions in dashboard
     - Desktop notification if update available
 <!-- - Added video example -->